Move BuildOperationState to its own top-level class
